requirements.txt
django
djangorestframework
pycryptodome